A large update
Succeeded in creating a system of wrappers to easily switch between persistent
and volatile based allocation/free modes utilizing intel pmdk

Have some simple results but they are too similar between pmem and non-pmem, I
will hopefully add an LSM tree next and start properly structuring a system of
different benchmark parameters.

it should be noted that my implementation of the libpmemobj is not technically
correct as it doesn't account for the possibility of virtual address changes
(since virtual addresses can change but the PMEMoid pointers won't). It hasn't
affected my tests thus far at the very least. It should also be noted that when
using the libpmemobj allocation functions, every block that is allocated
actually has an additional 64 bytes allocated with it (according to man pages)
which means that this further skews my results.