Another huge update
Added external LSM library and modified it to work with standardized functions
required by my testbench. Also fixed memory leaks, errors, and other silliness
pre-existing in the library. It passes simple tests without leaks now so
hopefully everything is ok. Performance is around expected range given current
tests. Kept the original source of the LSM tree as a submodule and will cite it
in the report.

Optimized vector: now stores capacity in first index and actual index is now
just computed by index + 1. This resulted in the vector being the fastest for
very basic tests. I imagine this outlook will change when the range of values
becomes larger and memory utilization of the vector likely will be the worst.

Fixed some simple things that were annoying me such as stuck to using uintptr_t
(mostly because it is possible to convert to it easily from (void *))

Next steps:
- actual benchmark (what I have now is simply a testbench)
- make LSM tree persistent (OH BOY, this is gonna be fun)